cmd/initContainer: Remove path on redirection when it exists
The redirectPath() function does not work correctly if containerPath
does not exist (like /media). The error message often looks like this:

Error: failed to redirect /media to /run/media: remove /media: no such file or directory

This makes redirectPath try to remove containerPath only if it exists.
The existence of containerPath is done using utils.PathExist().

https://github.com/containers/toolbox/pull/556

+++ src/cmd/initContainer.go
@@ -494,12 +494,14 @@ func redirectPath(containerPath, target string, folder bool) error {
 	logrus.Debugf("Preparing to redirect %s to %s", containerPath, target)
 	targetSanitized := sanitizeRedirectionTarget(target)
 
-	err := os.Remove(containerPath)
-	if folder {
-		if err != nil {
+	if utils.PathExists(containerPath) {
+		err := os.Remove(containerPath)
+		if err != nil && folder {
 			return fmt.Errorf("failed to redirect %s to %s: %w", containerPath, target, err)
 		}
+	}
 
+	if folder {
 		if err := os.MkdirAll(target, 0755); err != nil {
 			return fmt.Errorf("failed to redirect %s to %s: %w", containerPath, target, err)
 		}
